+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # 0.0.2
 
++ Implemented a STLC core language
+
 # 0.0.1
 
 + Create package on crates.io